# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Followup API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/request/tasks": {
            "get": {
                "tags": [
                    "request tasks"
                ],
                "summary": "Get all tasks",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            },
            "post": {
                "tags": [
                    "request tasks"
                ],
                "summary": "Add New task",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "po",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "request_form_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/tasks/user/requests": {
            "get": {
                "tags": [
                    "request tasks"
                ],
                "summary": "Get user tasks",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/tasks/accounts/reviews": {
            "get": {
                "tags": [
                    "request tasks"
                ],
                "summary": "Get accounts tasks reviews",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/tasks/accounts/{task}/change-status": {
            "post": {
                "tags": [
                    "request tasks"
                ],
                "summary": "account task change status",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "active",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/request/tasks/{task}/update": {
            "post": {
                "tags": [
                    "request tasks"
                ],
                "summary": "update task",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "po",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "request_form_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/request/tasks/{task}/destroy": {
            "post": {
                "tags": [
                    "request tasks"
                ],
                "summary": "update task",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
...
        "/vacations": {
            "get": {
                "tags": [
                    "vacations"
                ],
                "summary": "Get all vacations",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            },
            "post": {
                "tags": [
                    "vacations"
                ],
                "summary": "add New vacation",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "date_from",
                        "in": "formData",
                        "description": "2019-10-14",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "date_to",
                        "in": "formData",
                        "description": "2019-10-14",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
...
        "/users/vacations/manager-change-status/{vacation}": {
            "get": {
                "tags": [
                    "vacations"
                ],
                "summary": "Team Manager Change vacation Status ",
                "parameters": [
                    {
                        "name": "vacation",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "active",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/users/vacations/hr-change-status/{vacation}": {
            "get": {
                "tags": [
                    "vacations"
                ],
                "summary": "HR  Change vacation Status",
                "parameters": [
                    {
                        "name": "vacation",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "hr_active",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        }
    },
...
}